refactoring

+++ src/main/classes/controllers/CacheController.java
@@ -4,7 +4,9 @@ import caching.*;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mongodb.MongoClient;
+import com.mongodb.client.AggregateIterable;
 import com.mongodb.client.FindIterable;
+import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 
 
@@ -16,6 +18,7 @@ import org.bson.Document;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Logger;
 
 import static utils.DateUtil.*;
@@ -33,7 +36,7 @@ public class CacheController {
     private static final String CACHECONFIGS_COLLECTION_NAME = "cacheconfigs";
     private static final String CACHEKEYS_COLLECTION_NAME = "cachekeys";
     private final MongoDatabase mongoDatabase = new MongoClient().getDatabase(DATABASE_NAME);
-
+    private final Map<String, Cache> caches = new ConcurrentHashMap<>(new HashMap<>());
     private CacheRepository cacheRepository;
 
     public CacheController() throws CacheAlreadyExistsException, CacheNotFoundException, InvalidKeyException, IOException {
@@ -41,81 +44,117 @@ public class CacheController {
     }
     public Object create(CacheConfig cacheConfig) throws CacheAlreadyExistsException, JsonProcessingException { //DB
         String cacheName = cacheConfig.getName();
-        cacheRepository.addCache(new CacheImpl(cacheConfig));
+
+        Cache cache = caches.putIfAbsent(cacheName, new CacheImpl(cacheConfig));
+        if(cache != null) throw new CacheAlreadyExistsException(String.format("A cache with name: %s already exists. Try another cache name.", cacheName));
+
         mongoDatabase.getCollection(CACHECONFIGS_COLLECTION_NAME).insertOne(Document.parse(toJson(cacheConfig)));
         mongoDatabase.getCollection(CACHEKEYS_COLLECTION_NAME).insertOne(new Document("name", cacheName).append("keys", new ArrayList<>())); //<-- hmm
         return cacheConfig;
     }
     public Object edit(CacheConfig cacheConfig) throws CacheNotFoundException, JsonProcessingException { //DB
         String cacheName = cacheConfig.getName();
-        Cache cache = cacheRepository.getCache(cacheName);
-        mongoDatabase.getCollection(CACHECONFIGS_COLLECTION_NAME).findOneAndUpdate(new Document("name", cacheName), new Document("$set", Document.parse(toJson(cacheConfig))));
+        Cache cache = caches.computeIfPresent(cacheName, (k,c) -> {c.setCacheConfig(cacheConfig); return c;});
+        if(cache == null) throw new CacheNotFoundException(String.format("Cache with name: %s does not exist. You can create a new cache at route: /api/cache.", cacheName));
 
-        synchronized (cache){
-            cache.setCacheConfig(cacheConfig);
-        }
+        mongoDatabase.getCollection(CACHECONFIGS_COLLECTION_NAME).findOneAndUpdate(new Document("name", cacheName), new Document("$set", Document.parse(toJson(cacheConfig))));
         return cacheConfig;
     }
     public Object delete(String cacheName) throws CacheNotFoundException { //DB
+        Cache cache = caches.remove(cacheName);
+        if(cache == null) throw new CacheNotFoundException(String.format("Cache with name: %s does not exist. You can create a new cache at route: /api/cache.", cacheName));
+
         mongoDatabase.getCollection(CACHECONFIGS_COLLECTION_NAME).deleteOne(new Document("name", cacheName));
         mongoDatabase.getCollection(CACHEKEYS_COLLECTION_NAME).deleteOne(new Document("name", cacheName));
-        return cacheRepository.deleteCache(cacheName).getCacheConfig();
+        return cache.getCacheConfig();
     }
     public Object get(String cacheName) throws CacheNotFoundException {
-        return cacheRepository.getCache(cacheName).getCacheConfig();
+        Cache cache = caches.get(cacheName);
+        if(cache == null) throw new CacheNotFoundException(String.format("Cache with name: %s does not exist. You can create a new cache at route: /api/cache.", cacheName));
+        return cache.getCacheConfig();
     }
 
     public Object putKey(String cacheName, TreeMap<String,String> key) throws CacheNotFoundException, RequiresValidDateException, InvalidKeyException, JsonProcessingException, CacheAlreadyExistsException {
         LocalDateTime now = LocalDateTime.now();
         return putKey(cacheName, key, now);
     }
-    public Object putKey(String cacheName, TreeMap<String,String> key, LocalDateTime now) throws CacheNotFoundException, RequiresValidDateException, InvalidKeyException, JsonProcessingException, CacheAlreadyExistsException {
-        Cache cache = cacheRepository.getCache(cacheName);
-        if(cache.hasExpired(now)){ //lazy expire
-            logger.info(String.format("%s has expired and is reset.", cacheName));
-            CacheConfig cacheConfig = cache.getCacheConfig();
-            cacheConfig.setCreatedAt(now);
-            delete(cacheName);
-            create(cacheConfig);
-            cache = cacheRepository.getCache(cacheName);
-        }
+    public Object putKey(String cacheName, TreeMap<String,String> key, LocalDateTime now) throws CacheNotFoundException, JsonProcessingException {
+        //check key here instead?
+
+        Cache cache = caches.computeIfPresent(cacheName, (k,c) -> {
+            if(c.hasExpired(now)){
+                logger.info(String.format("%s has expired and is reset.", cacheName));
+                CacheConfig cacheConfig = c.getCacheConfig();
+                cacheConfig.setCreatedAt(now);
+                c = new CacheImpl(cacheConfig);
+            }
+            try {
+                c.put(key, now, 1);
+                return c;
+            } catch (InvalidKeyException e) {
+                throw new IllegalArgumentException(e.getMessage(), e);
+            }
+        });
+        if(cache == null)
+            throw new CacheNotFoundException(String.format("Cache with name: %s does not exist. You can create a new cache at route: /api/cache.", cacheName));
+
         Key keyObj = new Key(key, now);
         Document filter = new Document("name", cacheName);
         Document update = new Document("$push", new Document("keys", Document.parse(toJson(keyObj))));
         mongoDatabase.getCollection(CACHEKEYS_COLLECTION_NAME).updateOne(filter, update);
-        return cache.put(key, now, 1);
 
+        return cache.pointGet(key, now);
     }
     public Object getPointEntry(String cacheName, String date, TreeMap<String,String> key) throws RequiresValidDateException, CacheNotFoundException {
         validateISODate(date);
-        return cacheRepository.getCache(cacheName).pointGet(key, LocalDateTime.parse(date));
+        Cache cache = caches.get(cacheName);
+        if(cache == null)
+            throw new CacheNotFoundException(String.format("Cache with name: %s does not exist. You can create a new cache at route: /api/cache.", cacheName));
+
+        return cache.pointGet(key, LocalDateTime.parse(date));
     }
     public Object getPointsEntry(String cacheName, String startDate, String endDate, TreeMap<String,String> key) throws RequiresValidDateException, CacheNotFoundException {
         validateISODate(startDate);
         validateISODate(endDate);
-        return cacheRepository.getCache(cacheName).pointsGet(key, LocalDateTime.parse(startDate), LocalDateTime.parse(endDate));
+        Cache cache = caches.get(cacheName);
+        if(cache == null)
+            throw new CacheNotFoundException(String.format("Cache with name: %s does not exist. You can create a new cache at route: /api/cache.", cacheName));
+
+        return cache.pointsGet(key, LocalDateTime.parse(startDate), LocalDateTime.parse(endDate));
     }
 
     public Object getRangeEntry(String cacheName, String startDate, String endDate, TreeMap<String,String> key) throws RequiresValidDateException, CacheNotFoundException {
         validateISODate(startDate);
         validateISODate(endDate);
-        return cacheRepository.getCache(cacheName).rangeGet(key, LocalDateTime.parse(startDate), LocalDateTime.parse(endDate));
+        Cache cache = caches.get(cacheName);
+        if(cache == null)
+            throw new CacheNotFoundException(String.format("Cache with name: %s does not exist. You can create a new cache at route: /api/cache.", cacheName));
+
+        return cache.rangeGet(key, LocalDateTime.parse(startDate), LocalDateTime.parse(endDate));
     }
     public Object getTopEntry(String cacheName, int days) throws CacheNotFoundException {
-        return cacheRepository.getCache(cacheName).topGet(days);
+        Cache cache = caches.get(cacheName);
+        if(cache == null)
+            throw new CacheNotFoundException(String.format("Cache with name: %s does not exist. You can create a new cache at route: /api/cache.", cacheName));
+
+        return cache.topGet(days);
     }
 
     private void initialize() throws InvalidKeyException, IOException {
 
         FindIterable<Document> cacheConfigs = mongoDatabase.getCollection(CACHECONFIGS_COLLECTION_NAME).find();
-        Map<String, Cache> caches = new HashMap<>();
         ObjectMapper objectMapper = new ObjectMapper();
 
         for(Document dcc : cacheConfigs){
             CacheConfig cacheConfig = objectMapper.readValue(dcc.toJson(), CacheConfig.class);
             Cache cache = new CacheImpl(cacheConfig);
             String cacheName = cacheConfig.getName();
-            Document dcck = mongoDatabase.getCollection(CACHEKEYS_COLLECTION_NAME).find(new Document("name", cacheName)).first();
+
+            Document dcck = mongoDatabase.getCollection(CACHEKEYS_COLLECTION_NAME).aggregate(Arrays.asList(
+               new Document("$match", new Document("name", cacheName)),
+               new Document("$project", new Document("keys", new Document("$filter", new Document("input", "$keys").append("as", "key").append("cond", new Document("$gte", Arrays.asList("$$key.createdAt", toDate(cacheConfig.getCreatedAt())))))))
+            )).first();
+
             if(dcck != null){
                 @SuppressWarnings("unchecked")
                 List<Document> cacheKeys = (List<Document>)dcck.get("keys");
@@ -131,7 +170,6 @@ public class CacheController {
         }
         logger.info(String.format("Loaded %d caches.", caches.size()));
 
-        this.cacheRepository = new CacheRepository(caches);
     }
